reorganizing code for cleaner structure + addressing circular import
